+++ apps/web/src/types/react-table-config.d.ts
@@ -1,1 +1,6 @@
-export * from '@saas-ui-pro/react/src/data-grid/react-table-config.d'
+import { DataGridColumnMeta } from '@saas-ui-pro/react'
+
+declare module '@tanstack/react-table' {
+  interface ColumnMeta<TData, TValue>
+    extends DataGridColumnMeta<TData, TValue> {}
+}