+++ saas-ui/react/src/data-grid/data-grid.stories.tsx
@@ -16,6 +16,7 @@ import {
   Stack,
   TableRowProps,
   Tr,
+  VisuallyHidden,
 } from '@chakra-ui/react'
 import { rand, randFirstName, randUser } from '@ngneat/falso'
 import {
@@ -33,6 +34,8 @@ import { LuMoreVertical } from 'react-icons/lu'
 import {
   RiAddFill,
   RiArrowDownFill,
+  RiArrowLeftFill,
+  RiArrowRightFill,
   RiArrowUpFill,
   RiSubtractFill,
 } from 'react-icons/ri'
@@ -156,7 +159,7 @@ const columns: ColumnDef<ExampleData>[] = [
   },
   {
     id: 'action',
-    header: '',
+    header: () => <VisuallyHidden>Actions</VisuallyHidden>,
     cell: ActionCell,
     size: 50,
     enableSorting: false,
@@ -826,8 +829,12 @@ export const WithCustomIcons = {
           sortDescending: <RiArrowDownFill />,
           rowExpanded: <RiSubtractFill />,
           rowCollapsed: <RiAddFill />,
+          nextPage: <RiArrowRightFill />,
+          previousPage: <RiArrowLeftFill />,
         }}
-      />
+      >
+        <DataGridPagination />
+      </DataGrid>
     )
   },
 }
@@ -869,6 +876,59 @@ export const WithLargeDataSet = {
           },
         }}
       />
+    )
+  },
+}
+
+export const DynamicPagination = {
+  render() {
+    const ref = React.useRef<HTMLTableElement>(null)
+
+    const columns = React.useMemo(() => makeColumns(1_000), [])
+    const [data] = React.useState(makeVirtualizedData(1_000, columns))
+
+    const [pagination, setPagination] = React.useState({
+      pageIndex: 0,
+      pageSize: 20,
+    })
+
+    const calcPerPage = React.useCallback(() => {
+      const offset = 88 // header + footer (pagination)
+      const parent = ref.current?.parentElement
+      const gridHeight = parent?.offsetHeight ? parent.offsetHeight - offset : 0
+
+      const rowHeight =
+        parent?.querySelector<HTMLTableRowElement>('tbody > tr')
+          ?.offsetHeight ?? 40
+
+      if (gridHeight && rowHeight) {
+        setPagination((state) => ({
+          ...state,
+          pageSize: Math.ceil(gridHeight / rowHeight),
+        }))
+      }
+    }, [])
+
+    React.useEffect(() => {
+      calcPerPage()
+
+      window.addEventListener('resize', calcPerPage)
+
+      return () => window.removeEventListener('resize', calcPerPage)
+    }, [data])
+
+    return (
+      <DataGrid
+        ref={ref}
+        data={data}
+        columns={columns}
+        onPaginationChange={setPagination}
+        state={{
+          pagination,
+        }}
+      >
+        <DataGridPagination />
+      </DataGrid>
     )
   },
 }
@@ -925,7 +985,7 @@ export const UseColumns = {
         }),
         helper.display({
           id: 'action',
-          header: '',
+          header: 'Actions',
           cell: ActionCell,
           size: 50,
           enableSorting: false,