+++ packages/marketing/CHANGELOG.md
@@ -1,5 +1,28 @@
 # marketing
 
+## 6.0.0
+
+### Patch Changes
+
+- Updated dependencies [0ddcac0]
+- Updated dependencies [3ffd67e]
+- Updated dependencies [3ffd67e]
+- Updated dependencies [0ddcac0]
+  - @saas-ui-pro/react@0.35.0
+  - @app/config@11.0.0
+  - @ui/lib@11.0.0
+  - @ui/theme@11.0.0
+
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies [033b340]
+  - @saas-ui-pro/react@0.34.0
+  - @app/config@10.0.0
+  - @ui/lib@10.0.0
+  - @ui/theme@10.0.0
+
 ## 4.0.0
 
 ### Patch Changes